NMR-6922: rename NMRAxisLimits to AxisLimit, implemented by Axis directly
