+++ nmrfx-analyst-gui/src/main/java/org/nmrfx/processor/gui/spectra/NMRAxis.java
@@ -21,7 +21,7 @@ package org.nmrfx.processor.gui.spectra;
 import javafx.geometry.Orientation;
 import org.nmrfx.chart.Axis;
 
-public class NMRAxis extends Axis implements NMRAxisLimits {
+public class NMRAxis extends Axis {
     private boolean showTicsAndLabels = true;
 
     public NMRAxis(Orientation orientation, double lowerBound, double upperBound, double width, double height) {
@@ -30,11 +30,6 @@ public class NMRAxis extends Axis implements NMRAxisLimits {
 
     public void setShowTicsAndLabels(boolean state) {
         showTicsAndLabels = state;
-    }
-
-    @Override
-    public boolean getReverse() {
-        return isReversed();
     }
 
     public void updateStateAndLabel(String label) {
